mc13xxx-core: Kconfig: Config menu driven by specific IC type
This patch makes config items for the mc13892 and mc13783 distinct and
splits introduces a separate config item for spi interface support in
preparation for a separate i2c and spi backend.

The mc13xxx generic core is therefore selected by the spi (or i2c) item:
having it the other way around doesn't work for other drivers that will
need to distinguish between mc13783 and mc13892 (at build time).

Signed-off-by: Marc Reilly <marc@cpdesign.com.au>