input: mc13xxx_ts: add support for mc13892
The way the raw touchscreen co-ordinates are sampled have some minor
differences for the mc13892 and mc13873. This patch introduces an IC
specific handler to convert the raw samples into x, y, pressure readings
for the input sub-system.
The mc13892 ignores samples that have too high a contact resistance.
Behaviour for the mc13873 is (intended to be) preserved.

Help text for the config item changes to reflect support for both IC
types

Signed-off-by: Marc Reilly <marc@cpdesign.com.au>