mfd: mc13xxx: introduce battery and charger driver for mc13892
This driver provides basic functionality for the battery and (wall)
charger functions of the mc13892. (It should also work for mc13873 with
some small adjustments)
It uses a simple polling approach which checks the battery and charge
status every few seconds.

The battery coulomb counter (CC) is supported. Some CC information is kept
in the MEMORYA register of the device so that the battery level is
available across reset/power cycles.

Signed-off-by: Marc Reilly <marc@cpdesign.com.au>