power: mc13xxx_battery: add gpio controlled charging shunt
The mc13892 has a maximum (recommended) charger path current of 1200 mA,
but this amount includes supply current for the rest system as well as
the charging current for the battery. This means that systems with a
high quiescent power usage can have significantly increased recharge times.

To overcome this limitation, an additional current source can be connected
to Vbp - after the charge path sense resistor so it is not included in
the mc13892's charger current. (See MC13892 Appnotes)

This patch provides basic support for a gpio controlled fixed current
source. If available, the driver will enable the current shunt during
the constant current charging phase. An optional sense gpio provides
feedback that the shunt is in the correct state.
Note that the battery current sense can measure up to 3A, so up to 2.2A
shunt can be used and the coulomb counter will still operate correctly.

Signed-off-by: Marc Reilly <marc@cpdesign.com.au>