mc13xxx_battery: disable power limiting on charge path
After power-on using a discharged battery, the engagement of the power
limiting causes the charger to periodically be disconnected. The system
power drain is then too much to be held up by the (basically empty)
battery alone, and the battery drops below threshold and the system
powers off.

This change disables the power limiting. The FETs used for M1 and M2 have
such low Rdson that it would take over 5A to exceed device specs.

Signed-off-by: Marc Reilly <marc@cpdesign.com.au>