vpr200: decrease battery nominal capacity, allow specifying as param
This decreases the nominal battery capacity to 8Ah, which allows for
temperature and cycle de-rating. As a result, it provides a worst case
value for the current capacity so that the battery has _at least_ the
reported amount.

The nominal battery capacity can also be passed in as a boot parameter
from the bootloader. This is in anticipation of higher capacity batteries
in the future.

Signed-off-by: Marc Reilly <marc@cpdesign.com.au>