drivers: video: mxcfb: import the mxcfb driver from freescale BSP
This adds the minimal amount of the fb driver from the freescale BSP
so that the fsl gstreamer mmcodecs can be used. The codecs rely on
several ioctls.

Note that the Kconfig and Makefile are not tapped into the build system
at this point. (only new files are added in this commit).

Some changes from the original files were required to work with v3.2
kernel. I tried to mark these with // CPD comments. The dead code can be
removed later.

Signed-off-by: Marc Reilly <marc@cpdesign.com.au>