drivers: misc: add driver for BMP280
This adds a driver which is a partial backport of one from the 4.7 kernel,
but has the same module init/probe driver funcs like the old one. (The
newer driver uses the IIO subsystem which is not present in this kernel)

The functionality of this driver could/should be together with the bmp085
driver, but its not worth the extra effort.

Signed-off-by: Marc Reilly <marc@cpdesign.com.au>